Audio buffer in live transcription should be in float32 for vad filter to work #48

+++ src/whisper_ctranslate2/live.py
@@ -132,7 +132,7 @@ class Live:
                 )
 
             result = self.transcribe.inference(
-                audio=_buffer.flatten(),
+                audio=_buffer.flatten().astype("float32"),
                 task=self.task,
                 language=self.language,
                 verbose=self.verbose,
